Adding new class for handling the data, figure and pickle.
